8279565: riscv: RVB: Add byte reverse instruction
