+++ src/hotspot/cpu/riscv/riscv.ad
@@ -7558,11 +7558,11 @@ instruct bytes_reverse_int(iRegINoSp dst, iRegIorL2I src, rFlagsReg cr) %{
   match(Set dst (ReverseBytesI src));
   effect(TEMP cr);
 
-  ins_cost(ALU_COST * 17);
-  format %{ "grevw  $dst, $src\t#@bytes_reverse_int" %}
+  ins_cost(ALU_COST * 13);
+  format %{ "revb_w_w  $dst, $src\t#@bytes_reverse_int" %}
 
   ins_encode %{
-    __ grevw(as_Register($dst$$reg), as_Register($src$$reg));
+    __ revb_w_w(as_Register($dst$$reg), as_Register($src$$reg));
   %}
 
   ins_pipe(ialu_reg);
@@ -7572,11 +7572,11 @@ instruct bytes_reverse_long(iRegLNoSp dst, iRegL src, rFlagsReg cr) %{
   match(Set dst (ReverseBytesL src));
   effect(TEMP cr);
 
-  ins_cost(ALU_COST * 45);
-  format %{ "grev  $dst, $src\t#@bytes_reverse_long" %}
+  ins_cost(ALU_COST * 29);
+  format %{ "revb  $dst, $src\t#@bytes_reverse_long" %}
 
   ins_encode %{
-    __ grev(as_Register($dst$$reg), as_Register($src$$reg));
+    __ revb(as_Register($dst$$reg), as_Register($src$$reg));
   %}
 
   ins_pipe(ialu_reg);
@@ -7586,10 +7586,10 @@ instruct bytes_reverse_unsigned_short(iRegINoSp dst, iRegIorL2I src) %{
   match(Set dst (ReverseBytesUS src));
 
   ins_cost(ALU_COST * 5);
-  format %{ "grevhu  $dst, $src\t#@bytes_reverse_unsigned_short" %}
+  format %{ "revb_h_h_u  $dst, $src\t#@bytes_reverse_unsigned_short" %}
 
   ins_encode %{
-    __ grevhu(as_Register($dst$$reg), as_Register($src$$reg));
+    __ revb_h_h_u(as_Register($dst$$reg), as_Register($src$$reg));
   %}
 
   ins_pipe(ialu_reg);
@@ -7599,10 +7599,10 @@ instruct bytes_reverse_short(iRegINoSp dst, iRegIorL2I src) %{
   match(Set dst (ReverseBytesS src));
 
   ins_cost(ALU_COST * 5);
-  format %{ "grevh  $dst, $src\t#@bytes_reverse_short" %}
+  format %{ "revb_h_h  $dst, $src\t#@bytes_reverse_short" %}
 
   ins_encode %{
-    __ grevh(as_Register($dst$$reg), as_Register($src$$reg));
+    __ revb_h_h(as_Register($dst$$reg), as_Register($src$$reg));
   %}
 
   ins_pipe(ialu_reg);
